+++ HISTORY.rst
@@ -6,6 +6,7 @@ History
 Development
 +++++++++++
 - Fix form-check-{label,input} classes applied too broadly. These should only be applied to labels and inputs found underneath the enclosing widget div for radio and checkbox select, not on the whole document.
+- Allow to display radio and checkbox elements inline using the form-check classes.
 
 1.0.1 (2019-08-30)
 ++++++++++++++++++